Rename ReconnectProtocolFactory and ReconnectProtocolFactoryImpl as suggested
